rename generateFakerData to generateJSONData
